Generator of Manifest File with Barcode Support
Application with list of clients and their details which include
container(s) + item(s) based on barcodes scanned with the camera and
stored in the local database.
It has to create pdf file with all scanned container(s) + item(s), total
quantities of each as well as client details such as Name, Address,
Email, Phone, Comment, Collection Date which will be send stored or
typed email.